Refactor dealing with ObjectInfo in StorageObjectStorage
