+++ src/Storages/ObjectStorage/DataLakes/Iceberg/IcebergWrites.cpp
@@ -405,7 +405,7 @@ void generateManifestList(
             {
                 auto manifest_list = snapshots->getObject(static_cast<UInt32>(i))->getValue<String>(Iceberg::f_manifest_list);
 
-                StorageObjectStorage::ObjectInfo object_info(filename_generator.convertMetadataPathToStoragePath(manifest_list));
+                ObjectInfo object_info(filename_generator.convertMetadataPathToStoragePath(manifest_list));
                 auto manifest_list_buf = createReadBuffer(object_info, object_storage, context, getLogger("IcebergWrites"));
 
                 auto input_stream = std::make_unique<AvroInputStreamReadBufferAdapter>(*manifest_list_buf);
