+++ src/Storages/ObjectStorage/Iterators/ReadBufferIterator.h
@@ -1,17 +1,17 @@
 #pragma once
-#include <Interpreters/Context_fwd.h>
 #include <Formats/ReadSchemaUtils.h>
-#include <Storages/ObjectStorage/StorageObjectStorageSource.h>
+#include <Interpreters/Context_fwd.h>
+#include <Storages/ObjectStorage/Iterators/IObjectIterator.h>
+#include <Storages/ObjectStorage/StorageObjectStorageConfiguration.h>
 
 
 namespace DB
 {
 
+
 class ReadBufferIterator : public IReadBufferIterator, WithContext
 {
 public:
-    using ObjectInfos = StorageObjectStorage::ObjectInfos;
-
     ReadBufferIterator(
         ObjectStoragePtr object_storage_,
         StorageObjectStorageConfigurationPtr configuration_,
