+++ src/Storages/ObjectStorage/StorageObjectStorageSource.cpp
@@ -1,16 +1,21 @@
 #include <memory>
 #include <optional>
-#include <Common/SipHash.h>
 #include <Core/Settings.h>
 #include <Disks/IO/AsynchronousBoundedReadBuffer.h>
+#include <Disks/IO/CachedOnDiskReadBufferFromFile.h>
+#include <Disks/ObjectStorages/IObjectStorage.h>
 #include <Disks/ObjectStorages/ObjectStorageIterator.h>
 #include <Formats/FormatFactory.h>
 #include <Formats/ReadSchemaUtils.h>
+#include <IO/Archives/ArchiveUtils.h>
 #include <IO/Archives/createArchiveReader.h>
 #include <IO/ReadBufferFromFileBase.h>
-#include <IO/Archives/ArchiveUtils.h>
+#include <Interpreters/Cache/FileCache.h>
 #include <Interpreters/Cache/FileCacheFactory.h>
+#include <Interpreters/Cache/FileCacheKey.h>
+#include <Interpreters/Context.h>
 #include <Interpreters/ExpressionActions.h>
+#include <Interpreters/convertFieldToType.h>
 #include <Processors/Executors/PullingPipelineExecutor.h>
 #include <Processors/Sources/ConstChunkGenerator.h>
 #include <Processors/Transforms/AddingDefaultsTransform.h>
@@ -18,18 +23,13 @@
 #include <Processors/Transforms/ExtractColumnsTransform.h>
 #include <QueryPipeline/QueryPipelineBuilder.h>
 #include <Storages/Cache/SchemaCache.h>
+#include <Storages/HivePartitioningUtils.h>
+#include <Storages/ObjectStorage/DataLakes/DataLakeConfiguration.h>
 #include <Storages/ObjectStorage/StorageObjectStorage.h>
 #include <Storages/ObjectStorage/StorageObjectStorageSource.h>
-#include <Storages/ObjectStorage/DataLakes/DataLakeConfiguration.h>
 #include <Storages/VirtualColumnUtils.h>
-#include <Storages/HivePartitioningUtils.h>
+#include <Common/SipHash.h>
 #include <Common/parseGlobs.h>
-#include <Disks/IO/CachedOnDiskReadBufferFromFile.h>
-#include <Disks/ObjectStorages/IObjectStorage.h>
-#include <Interpreters/Cache/FileCache.h>
-#include <Interpreters/Cache/FileCacheKey.h>
-#include <Interpreters/convertFieldToType.h>
-#include <Interpreters/Context.h>
 #if ENABLE_DISTRIBUTED_CACHE
 #include <DistributedCache/DistributedCacheRegistry.h>
 #include <Disks/IO/ReadBufferFromDistributedCache.h>
@@ -113,20 +113,6 @@ StorageObjectStorageSource::StorageObjectStorageSource(
 StorageObjectStorageSource::~StorageObjectStorageSource()
 {
     create_reader_pool->wait();
-}
-
-std::string StorageObjectStorageSource::getUniqueStoragePathIdentifier(
-    const StorageObjectStorageConfiguration & configuration,
-    const ObjectInfo & object_info,
-    bool include_connection_info)
-{
-    auto path = object_info.getPath();
-    if (path.starts_with("/"))
-        path = path.substr(1);
-
-    if (include_connection_info)
-        return fs::path(configuration.getDataSourceDescription()) / path;
-    return fs::path(configuration.getNamespace()) / path;
 }
 
 std::shared_ptr<IObjectIterator> StorageObjectStorageSource::createFileIterator(
@@ -681,7 +667,7 @@ size_t StorageObjectStorageSource::GlobIterator::estimatedKeysCount()
     return object_infos.size();
 }
 
-StorageObjectStorage::ObjectInfoPtr StorageObjectStorageSource::GlobIterator::next(size_t processor)
+ObjectInfoPtr StorageObjectStorageSource::GlobIterator::next(size_t processor)
 {
     std::lock_guard lock(next_mutex);
     auto object_info = nextUnlocked(processor);
@@ -695,7 +681,7 @@ StorageObjectStorage::ObjectInfoPtr StorageObjectStorageSource::GlobIterator::ne
     return object_info;
 }
 
-StorageObjectStorage::ObjectInfoPtr StorageObjectStorageSource::GlobIterator::nextUnlocked(size_t /* processor */)
+ObjectInfoPtr StorageObjectStorageSource::GlobIterator::nextUnlocked(size_t /* processor */)
 {
     bool current_batch_processed = object_infos.empty() || index >= object_infos.size();
     if (is_finished && current_batch_processed)
@@ -789,7 +775,7 @@ StorageObjectStorageSource::KeysIterator::KeysIterator(
     }
 }
 
-StorageObjectStorage::ObjectInfoPtr StorageObjectStorageSource::KeysIterator::next(size_t /* processor */)
+ObjectInfoPtr StorageObjectStorageSource::KeysIterator::next(size_t /* processor */)
 {
     while (true)
     {
@@ -886,7 +872,7 @@ StorageObjectStorageSource::ReadTaskIterator::ReadTaskIterator(
     }
 }
 
-StorageObjectStorage::ObjectInfoPtr StorageObjectStorageSource::ReadTaskIterator::next(size_t)
+ObjectInfoPtr StorageObjectStorageSource::ReadTaskIterator::next(size_t)
 {
     size_t current_index = index.fetch_add(1, std::memory_order_relaxed);
 
@@ -941,125 +927,6 @@ ObjectInfoPtr StorageObjectStorageSource::ReadTaskIterator::createObjectInfoInAr
 
     return std::make_shared<ArchiveIterator::ObjectInfoInArchive>(
         archive_object, path_in_archive, archive_reader, archive_reader->getFileInfo(path_in_archive));
-}
-
-static IArchiveReader::NameFilter createArchivePathFilter(const std::string & archive_pattern)
-{
-    auto matcher = std::make_shared<re2::RE2>(makeRegexpPatternFromGlobs(archive_pattern));
-    if (!matcher->ok())
-    {
-        throw Exception(ErrorCodes::CANNOT_COMPILE_REGEXP,
-                        "Cannot compile regex from glob ({}): {}",
-                        archive_pattern, matcher->error());
-    }
-    return [matcher](const std::string & p) mutable { return re2::RE2::FullMatch(p, *matcher); };
-}
-
-StorageObjectStorageSource::ArchiveIterator::ObjectInfoInArchive::ObjectInfoInArchive(
-    ObjectInfoPtr archive_object_,
-    const std::string & path_in_archive_,
-    std::shared_ptr<IArchiveReader> archive_reader_,
-    IArchiveReader::FileInfo && file_info_)
-    : archive_object(archive_object_), path_in_archive(path_in_archive_), archive_reader(archive_reader_), file_info(file_info_)
-{
-}
-
-StorageObjectStorageSource::ArchiveIterator::ArchiveIterator(
-    ObjectStoragePtr object_storage_,
-    StorageObjectStorageConfigurationPtr configuration_,
-    std::unique_ptr<IObjectIterator> archives_iterator_,
-    ContextPtr context_,
-    ObjectInfos * read_keys_,
-    bool ignore_archive_globs_)
-    : WithContext(context_)
-    , object_storage(object_storage_)
-    , is_path_in_archive_with_globs(configuration_->isPathInArchiveWithGlobs())
-    , archives_iterator(std::move(archives_iterator_))
-    , filter(is_path_in_archive_with_globs ? createArchivePathFilter(configuration_->getPathInArchive()) : IArchiveReader::NameFilter{})
-    , log(getLogger("ArchiveIterator"))
-    , path_in_archive(is_path_in_archive_with_globs ? "" : configuration_->getPathInArchive())
-    , read_keys(read_keys_)
-    , ignore_archive_globs(ignore_archive_globs_)
-{
-}
-
-std::shared_ptr<IArchiveReader>
-StorageObjectStorageSource::ArchiveIterator::createArchiveReader(ObjectInfoPtr object_info) const
-{
-    const auto size = object_info->metadata->size_bytes;
-    return DB::createArchiveReader(
-        /* path_to_archive */
-        object_info->getPath(),
-        /* archive_read_function */ [=, this]() { return createReadBuffer(*object_info, object_storage, getContext(), log); },
-        /* archive_size */ size);
-}
-
-ObjectInfoPtr StorageObjectStorageSource::ArchiveIterator::next(size_t processor)
-{
-    std::unique_lock lock{next_mutex};
-    IArchiveReader::FileInfo current_file_info{};
-    while (true)
-    {
-        if (filter)
-        {
-            if (!file_enumerator)
-            {
-                archive_object = archives_iterator->next(processor);
-                if (!archive_object)
-                {
-                    LOG_TEST(log, "Archives are processed");
-                    return {};
-                }
-
-                if (!archive_object->metadata)
-                    archive_object->metadata = object_storage->getObjectMetadata(archive_object->getPath());
-
-                archive_reader = createArchiveReader(archive_object);
-                file_enumerator = archive_reader->firstFile();
-                if (!file_enumerator)
-                    continue;
-            }
-            else if (!file_enumerator->nextFile() || ignore_archive_globs)
-            {
-                file_enumerator.reset();
-                continue;
-            }
-
-            path_in_archive = file_enumerator->getFileName();
-            LOG_TEST(log, "Path in archive: {}", path_in_archive);
-            if (!filter(path_in_archive))
-                continue;
-            current_file_info = file_enumerator->getFileInfo();
-        }
-        else
-        {
-            archive_object = archives_iterator->next(processor);
-            if (!archive_object)
-                return {};
-
-            if (!archive_object->metadata)
-                archive_object->metadata = object_storage->getObjectMetadata(archive_object->getPath());
-
-            archive_reader = createArchiveReader(archive_object);
-            if (!archive_reader->fileExists(path_in_archive))
-                continue;
-            current_file_info = archive_reader->getFileInfo(path_in_archive);
-        }
-        break;
-    }
-
-    auto object_in_archive
-        = std::make_shared<ObjectInfoInArchive>(archive_object, path_in_archive, archive_reader, std::move(current_file_info));
-
-    if (read_keys != nullptr)
-        read_keys->push_back(object_in_archive);
-
-    return object_in_archive;
-}
-
-size_t StorageObjectStorageSource::ArchiveIterator::estimatedKeysCount()
-{
-    return archives_iterator->estimatedKeysCount();
 }
 
 }
