+++ src/Storages/ObjectStorage/Utils.h
@@ -32,4 +32,7 @@ std::unique_ptr<ReadBufferFromFileBase> createReadBuffer(
     const ContextPtr & context_,
     const LoggerPtr & log,
     const std::optional<ReadSettings> & read_settings = std::nullopt);
+
+std::string getUniqueStoragePathIdentifier(
+    const StorageObjectStorageConfiguration & configuration, const ObjectInfo & object_info, bool include_connection_info = true);
 }
